requirements.txt
numpy==2.2.4
opencv-python==4.11.0.86
paho-mqtt==2.1.0
python-dotenv==1.1.0
